# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nova-post/cities": {
      "get": {
        "description": "Case-insensitive search for Nova Post cities/settlements by name.",
        "operationId": "NovaPostController_getCities",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search cities",
        "tags": [
          "/nova-post"
        ]
      }
    },
    "/nova-post/warehouses": {
      "get": {
        "description": "Returns Nova Post warehouses for the given city ref. Optional `type` filter: PARCEL_LOCKER (поштомат), POST (поштове відділення), CARGO (вантажне відділення). Optional `q` (same idea as on `/nova-post/cities`): when non-empty after trim, returns only branches where the substring appears in the warehouse number (string form), `description`, or `shortAddress` — case-insensitive for text, with whitespace normalized in the query and flexible matching between words. When `q` is omitted or only whitespace, returns the full list for `cityRef` + `type` as before.",
        "operationId": "NovaPostController_getWarehouses",
        "parameters": [
          {
            "name": "cityRef",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PARCEL_LOCKER",
                "POST",
                "CARGO"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Optional search: warehouse number (substring on numeric №) and/or fragment of description / short address. Case-insensitive for text; trim and repeated spaces collapsed; word gaps match flexibly. Omit or leave blank for all warehouses in the city (and type).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get warehouses by city",
        "tags": [
          "/nova-post"
        ]
      }
    },
...
    "/discount-coupons": {
      "get": {
        "description": "Get paginated discount coupons with optional filters by active flag and code search query. Admin only.",
        "operationId": "DiscountCouponController_findAll",
        "parameters": [
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Filter by active flag",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by coupon code substring",
            "schema": {
              "example": "ZY64",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all discount coupons",
        "tags": [
          "/discount-coupons"
        ]
      },
      "post": {
        "description": "Create a discount coupon with internal number in DIS-0000123 format and random 10-character code (A-Z, 0-9). Admin only.",
        "operationId": "DiscountCouponController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscountCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create discount coupon",
        "tags": [
          "/discount-coupons"
        ]
      }
    },
    "/discount-coupons/{id}": {
      "get": {
        "description": "Get discount coupon details by MongoDB id. Admin only.",
        "operationId": "DiscountCouponController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get discount coupon by id",
        "tags": [
          "/discount-coupons"
        ]
      },
      "patch": {
        "description": "Update discount coupon fields (discount percent, expiration, active flag). Admin only.",
        "operationId": "DiscountCouponController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiscountCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update discount coupon",
        "tags": [
          "/discount-coupons"
        ]
      },
      "delete": {
        "description": "Delete discount coupon by id. Admin only.",
        "operationId": "DiscountCouponController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete discount coupon",
        "tags": [
          "/discount-coupons"
        ]
      }
    },
    "/discount-coupons/validate": {
      "post": {
        "description": "Validate coupon code for checkout. Returns validity, and for valid coupons returns discount details.",
        "operationId": "DiscountCouponController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDiscountCouponDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Validate discount coupon",
        "tags": [
          "/discount-coupons"
        ]
      }
    },
...
  },
  "info": {
    "title": "Urban Tab API",
    "description": "API документація",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateDiscountCouponDto": {
        "type": "object",
        "properties": {
          "discount_percent": {
            "type": "number",
            "example": 15,
            "minimum": 0,
            "maximum": 100
          },
          "valid_until": {
            "type": "string",
            "example": "2026-12-31T23:59:59.000Z",
            "description": "Coupon expiration datetime in ISO format"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "discount_percent",
          "valid_until"
        ]
      },
      "ValidateDiscountCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ZY64GM08WT"
          }
        },
        "required": [
          "code"
        ]
      },
      "UpdateDiscountCouponDto": {
        "type": "object",
        "properties": {
          "discount_percent": {
            "type": "number",
            "example": 15,
            "minimum": 0,
            "maximum": 100
          },
          "valid_until": {
            "type": "string",
            "example": "2026-12-31T23:59:59.000Z",
            "description": "Coupon expiration datetime in ISO format"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        }
      },
...
    }
  }
}